fix the tests to float64

+++ tests/test_inference.py
@@ -5,12 +5,12 @@ from grf_gp.utils.sparse_lo import SparseLinearOperator
 
 
 def test_pathwise_conditioning_smoke_shape():
-    eye = torch.eye(4)
+    eye = torch.eye(4, dtype=torch.float64)
     phi = 1 * SparseLinearOperator(eye.to_sparse_csr())
     x_train = torch.tensor([0, 1], dtype=torch.int64)
     x_test = torch.tensor([2, 3], dtype=torch.int64)
-    y_train = torch.tensor([0.5, -0.5])
-    noise_std = torch.tensor(0.1)
+    y_train = torch.tensor([0.5, -0.5], dtype=torch.float64)
+    noise_std = torch.tensor(0.1, dtype=torch.float64)
 
     samples = pathwise_conditioning(
         x_train=x_train,
